Migrate formatToOrdinal over into utils/formatters
