+++ src/utils/formatters.js
@@ -106,6 +106,13 @@ export function formatDateTime(date) {
   return undefined;
 }
 
+// maps int to int with ordinal 1 -> 1st, 2 -> 2nd, 3rd ...
+export const formatToOrdinal = (n) => {
+  const s = ['th', 'st', 'nd', 'rd'];
+  const v = n % 100;
+  return n + (s[(v - 20) % 10] || s[v] || s[0]);
+};
+
 // Format orders type (ex: PERMANENT_CHANGE_OF_STATION => Permanent change of station)
 export function formatOrderType(orderType) {
   return orderType
